added user factory and separated the factories directives and controllers
