+++ scotch-chat-main-app/app/public/js/directives/ngEnter-directive.js
@@ -1,0 +1,14 @@
+//ng-enter directive
+app.directive('ngEnter', function () {
+    return function (scope, element, attrs) {
+        element.bind("keydown keypress", function (event) {
+            if (event.which === 13) {
+                scope.$apply(function () {
+                    scope.$eval(attrs.ngEnter);
+                });
+
+                event.preventDefault();
+            }
+        });
+    };
+});
