+++ scotch-chat-main-app/app/public/js/factories/gui-factory.js
@@ -1,0 +1,8 @@
+//Services to interact with nodewebkit GUI and Window
+app.factory('GUI', function () {
+    //Return nw.gui
+    return require('nw.gui');
+});
+app.factory('Window', function (GUI) {
+    return GUI.Window.get();
+});
