+++ scotch-chat-main-app/app/public/js/factories/socket-factory.js
@@ -1,0 +1,11 @@
+
+//Service to interact with the socket library
+app.factory('socket', function (socketFactory) {
+    var myIoSocket = io.connect(serverBaseUrl);
+
+    var socket = socketFactory({
+        ioSocket: myIoSocket
+    });
+
+    return socket;
+});
