+++ scotch-chat-main-app/app/views/index.html
@@ -17,6 +17,12 @@
     <script type="text/javascript" src="libs/angular-material-icons/angular-material-icons.js"></script>
 
     <script src="js/app.js"></script>
+    <script src="js/factories/user-factory.js"></script>
+    <script src="js/factories/gui-factory.js"></script>
+    <script src="js/controllers/main-ctrl.js"></script>
+    <script src="js/factories/socket-factory.js"></script>
+    <script src="js/directives/ngEnter-directive.js"></script>
+    <script src="js/controllers/username-dialog-ctrl.js"></script>
 </head>
 
 <body>
@@ -24,7 +30,21 @@
     <div class='col-xs-6'>
     <div class="canvas-container"><canvas id="my-canvas" width="400" height="600">I am canvas</canvas></div>
     <button onclick="getInfo()">log</button>
-
+    <div class="tools">
+  <a href="#colors_sketch" data-download="png" style="float: right; width: 100px;">Download</a>
+</div>
+    <div class="canvas-container"><canvas id="my-canvas" id="color-sketch" width="400" height="600">I am canvas</canvas></div>
+    <script type="text/javascript">
+    $(function() {
+        $.each(['#f00', '#ff0', '#0f0', '#0ff', '#00f', '#f0f', '#000', '#fff'], function() {
+        $('#colors_demo .tools').append("<a href='#colors_sketch' data-color='" + this + "' style='width: 10px; background: " + this + ";'></a> ");
+        });
+        $.each([3, 5, 10, 15], function() {
+        $('#colors_demo .tools').append("<a href='#colors_sketch' data-size='" + this + "' style='background: #ccc'>" + this + "</a> ");
+        });
+        $('#colors_sketch').sketch();
+    });
+    </script>
 </div>
     <div class='col-xs-6'>
     <md-content>
@@ -129,6 +149,8 @@
                 context.lineTo(evt.layerX+1, evt.layerY+1);
                 context.stroke();
             }
-        }, false);    
+        }, false);   
+        
+        var __slice = Array.prototype.slice;
     </script>
 </html>
